requirements.txt
Django==1.10.5
django-environ==0.4.1
django-extensions==1.7.4
sorl-thumbnail==12.3
django-gravatar2==1.4.0
django-pipeline==1.6.9
requests==2.11.1
-e git+https://github.com/pythonbrasil/django-bootstrap-toolkit.git@v1-associados#egg=django_bootstrap_toolkit-dev
dj-database-url
psycopg2==2.6.2
gunicorn==19.6.0
lxml==3.6.4
#-e git+https://github.com/pythonbrasil/django-s3-folder-storage.git@v1-associados#egg=django_s3_folder_storage-dev
-e git+https://github.com/pythonbrasil/django-s3-folder-storage.git@d7e8b7939b554e8fdc6815da5fa349bf7d5d8aaa#egg=django_s3_folder_storage
slumber==0.7.1
boto==2.42.0
django-storages==1.5.1
six==1.10.0
raven==5.27.1
django-localflavor==1.3
cssselect==0.9.2
#django-municipios==0.8.1
-e git+https://github.com/lskbr/django-municipios.git#egg=django-municipios

